THE PLAN IS SIMPLEEEEEEEEEEEEEEEEEE
The goal is to make 40 odd classes using these 3. Whack in the OCP shite we did in the lab (create interfaces, etc.) Interface inheritance is easier apparently so prob use that. Make your own branch and DO NOT commit to master. Look up how GitHub works and please do not overwrite work you or others have done. I lost all my python shit by accident by messing up which branch i was committing to. I'm just gonna make a branch with my name on it so probably just do the same. Work on what you are working on but don't commit to master.